Accept null values & Fix function arguments data types requirements

+++ structures/methods/drop.js
@@ -3,7 +3,7 @@
 const errors = require('../errors/strings.js');
 
 module.exports = async function(table){
-	if(!table) throw new TypeError(errors.table.replace("{received}", table));
+	if(!table || typeof table !== "string") throw new TypeError(errors.table.replace("{received}", typeof table));
 	
 	this.db.query("DROP TABLE " + table);
 	return true;
